Minor corrections

+++ Sources/Contact/Contact.cpp
@@ -32,7 +32,7 @@ void Contact::TimeStepInitialization() noexcept {
 }
 
 void Contact::inFile(FILE *ft) noexcept {
-    fprintf(ft,"Contact : type = %d\n",type);
+    fprintf(ft,"Contact : type = %d\n",static_cast<int>(type));
 	switch (type) {
 		case Type::SphereSphere:
 			if(sa) fprintf(ft,"Sph : %d\n",sa->Num());
@@ -49,7 +49,7 @@ void Contact::inFile(FILE *ft) noexcept {
 }
 
 void Contact::Display() const noexcept {
-	printf("Contact : type = %d\n",type);
+	printf("Contact : type = %d\n",static_cast<int>(type));
 	switch (type) {
 		case Type::SphereSphere:
 			if(sa) printf("Sph : %d\n",sa->Num());
@@ -67,43 +67,43 @@ void Contact::Display() const noexcept {
 }
 
 void Contact::WriteOutFile(FILE *ft,double time) const noexcept {
-	fprintf(ft,"%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,delta,xi,yi,zi,xf,yf,zf);
+	fprintf(ft,"%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),delta,xi,yi,zi,xf,yf,zf);
 }
 
 void Contact::WriteOutFileDetails(FILE *ft,double time) const noexcept {
 	 switch(type){
 	  	case Type::SphereSphere: //0
-   fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),sb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+   fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),sb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::SphereBody: //10
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),bb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),bb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::BodyBody: //20
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,ba->Num(),bb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),ba->Num(),bb->Num(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::SpherePlan: //1
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),pa->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),pa->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::BodyPlan: //11
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,ba->Num(),pa->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),ba->Num(),pa->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::SpherePlanR: //2
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),par->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),par->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::BodyPlanR: //12
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,ba->Num(),par->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),ba->Num(),par->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::SphereCone: //3
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),cn->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),cn->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::BodyCone: //12
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,ba->Num(),cn->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),ba->Num(),cn->Numero(),px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::SphereElbow: //4
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,sa->Num(),ew->numero,px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),sa->Num(),ew->numero,px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		case Type::BodyElbow: //14
-	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",type,ba->Num(),ew->numero,px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
+	 fprintf(ft,"%d\t%d\t%d\t%e\t%e\t%e\t%e\t%d\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\t%e\n",static_cast<int>(type),ba->Num(),ew->numero,px,py,pz,delta,Stick,nx,ny,nz,tx,ty,tz,Fx,Fy,Fz);
 	 break;
 		default:
 			break;
